moved split_view from PGC into ToB #22600
