+++ learning/tour-of-beam/frontend/lib/components/split_view/pan.dart
