+++ playground/frontend/playground_components/pubspec.yaml
@@ -31,7 +31,6 @@ dependencies:
   google_fonts: ^3.0.1
   provider: ^6.0.3
   shared_preferences: ^2.0.15
-  split_view: ^3.2.1
 
 dev_dependencies:
   build_runner: ^2.2.0
